+++ moor/CHANGELOG.md
@@ -4,6 +4,7 @@
   for details.
 - Add `CASE WHEN` expressions with the `caseMatch` method on `Expression`
 - Add `insertReturning` which returns the generated row after an insert
+- Add `min`, `max` and `avg` functions for `DateTime` expressions
 - On supported platforms, cancel pending stream selects when the stream is disposed
   - `moor_flutter` is supported
   - `moor/ffi` is supported when used on a background isolate